CRUD Usuario  - Queda por hacer el file upload y mejorar las validaciones

+++ src/main/java/com/qveo/qveoweb/dao/UsuarioDao.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UsuarioDao extends JpaRepository<Usuario, Long> {
+public interface UsuarioDao extends JpaRepository<Usuario, Integer> {
 
     List<Usuario> findAll();
 }
